feature(notifications): add notifications
 - add view notification feature
 - add in app notificatiions
 - add email notifications
 - refactor views for articles

 [Finishes #161966618]
