+++ authors/apps/articles/templates/notification_email.html
@@ -1,0 +1,6 @@
+<h1>Hello {{ user }}</h1>
+<p>You have a new notification</p>
+<p>{{ author }} has published a new article Titled - {{ title }}</p>
+<p>Login to your account to view the Article.</p>
+<p>Click on the link to login</p>
+<p><a href="{{ link }}">Login To Authors Haven</a></p>
